Fix \Drupal\Tests\views\Functional\Plugin\DisplayFeedTranslationTest

+++ core/modules/path_alias/src/AliasManager.php
@@ -263,7 +263,9 @@ class AliasManager implements AliasManagerInterface {
     $this->noAlias = [];
     $this->langcodePreloaded = [];
     $this->preloadedPathLookups = [];
-    $this->cache->delete($this->cacheKey);
+    if (!empty($this->cacheKey)) {
+      $this->cache->delete($this->cacheKey);
+    }
     $this->pathAliasWhitelistRebuild($source);
   }
